new algoritm

+++ Fillit/print.c
@@ -6,20 +6,20 @@
 /*   By: mdirect <mdirect@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/31 13:10:30 by mdirect           #+#    #+#             */
-/*   Updated: 2019/11/16 21:45:38 by mdirect          ###   ########.fr       */
+/*   Updated: 2019/11/20 13:15:07 by mdirect          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fillit.h"
 
-void	ft_print_bit(uint64_t n, int y)
+void	ft_print_bit(uint128_t n, int y)
 {
-	uint64_t i;
+	int i;
 
 	i = y * y - 1;
 	while (i != UINT64_MAX)
 	{
-		if (n & ((uint64_t)1 << i))
+		if (n & ((uint128_t)1 << i))
 			write(1, "#", 1);
 		else
 			write(1, ".", 1);
@@ -30,11 +30,11 @@ void	ft_print_bit(uint64_t n, int y)
 	write(1, "\n", 1);
 }
 
-void	ft_print_map(uint64_t *tetra, int y)
+void	ft_print_map(uint128_t *tetra, int y)
 {
-	uint64_t i;
-	int		j;
-	char	letter;
+	int			i;
+	int			j;
+	char		letter;
 
 	i = y * y - 1;
 	while (i != UINT64_MAX)
@@ -42,7 +42,7 @@ void	ft_print_map(uint64_t *tetra, int y)
 		j = 0;
 		while (tetra[j])
 		{
-			if ((tetra[j] & ((uint64_t)1 << i)))
+			if ((tetra[j] & ((uint128_t)1 << i)))
 			{
 				letter = j + 65;
 				write(1, &letter, 1);
